+++ include/ambition/7kaaInterface/input.hh
@@ -44,6 +44,8 @@ enum class ReportType {
   TradeFirms,
   MilitaryTroops,
   MilitaryUnits,
+  TechnologyTechnologies,
+  TechnologyGods,
   News,
 };